Fix arg type

+++ best.py
@@ -76,7 +76,7 @@ def load_config(ctx, value):
 
 
 @click.command(help='Best fitness value.')
-@click.option('-m', '--float-max', type=int, default=float_info.max, help='Be quieter.')
+@click.option('-m', '--float-max', type=float, default=float_info.max, help='Be quieter.')
 @click.option('-q', '--quiet', count=True, help='Be quieter.')
 @click.option('-v', '--verbose', count=True, help='Be more verbose.')
 @click.option('-c', '--config',
